Simplify SandboxTestRunner.
SandboxTestRunner is now abstract, and only concerned with running tests in a sandbox provided by a concrete class.
